ensemble_app/requirements.txt
Django==4.2.16
pandas==2.2.3
openpyxl==3.1.5
django-autocomplete-light==3.12.1
num2words==0.5.13
psycopg2-binary==2.9.9
gunicorn==23.0.0
python-decouple==3.8
python-dateutil==2.9.0
redis==5.1.1
django-widget-tweaks==1.5.0
python-docx
django-mssql-backend 
pyodbc
celery[redis]
requests